docs(custom operators): Add arrayEqual and arrayNotEqual
The `arrayEqual` and `arrayNotEqual` operators are used to compare a fact that is an array of one element to a value. This is useful when you want to compare a fact that can be an array containing a single element or a single value to a value. These operators are not part of the json-rules-engine library and have been added to the quoting-service to maintain the legacy behavior of the json-rules-engine up to v5.0.2